refactor(hooks): split useGameDetail into focused hooks
Replace the 673-line useGameDetail aggregator with four single-domain
hooks: useGameMeta (game/members/GM actions), useAvailability,
useSessions, and usePlayDates. The supabase client is now hoisted to
module scope so it has stable identity, eliminating the dozen
"supabase is stable" eslint disables that papered over the dependency
warnings.

Cross-coupling (e.g., removePlayer also clearing that user's data
from the suggestions array) is now explicit at the call site:
useAvailability exposes a removePlayerData() coordination hook that
the page invokes after a successful removePlayer.

Adds withOptimistic() helper + tests for the optimistic-update
recipe (apply → mutate → revert-on-error / reconcile-on-success).
